#31 Rename quantities directory to factories
