+++ statistical_modeling/__init__.py
@@ -1,8 +1,8 @@
 from .sample import *
 from .distributions.base import *
-from .quantities.mean import Mean
-from .quantities.variance import Variance
-from .quantities.acf import ACF
+from .factories.mean import Mean
+from .factories.variance import Variance
+from .factories.acf import ACF
 
 from .distributions import (binomial,
                             chi_square,
